backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
beautifulsoup4==4.12.2
lxml==6.0.0
python-docx==1.2.0
